Adopt a strict warning set; firmware is clean under it
Now that a compiler exists it is worth asking it harder questions than
-Wall. The build now uses -Wall -Wextra -Wconversion -Wsign-compare
-Wshadow, and the codebase is clean under all of it.

-Wconversion is the one that earns its place here: in register-level
code it is what catches a mask or shift that silently truncates on its
way into a hardware register. It found exactly three hits across 5,400
lines, and all three were benign integer-promotion artefacts rather than
real defects - PCR_AF's result is bounded to 3 << 10 by its own mask,
and the CJ125/L9779 fault decoders mask to 2 bits. Each is now written
with an explicit cast that documents why it is lossless, rather than
suppressed, so the warning stays usable and anything it reports from
here is worth reading.

Only expected output is the two documented unused-function warnings for
injection.c's us_to_ticks/angle_to_ticks, which wait on the crank
trigger wheel's tooth count.

+++ ecu-firmware/inc/siul2.h
@@ -52,7 +52,11 @@
  *   0 = GPIO, 1/2/3 = alternate function 1/2/3 - WHICH real peripheral
  *   signal each AF number maps to is per-pin, from Table 4-1
  *   ("Functional port pins") - not encoded here, see the TODO below. */
-#define PCR_AF(n)    (((uint32_t)(n) & 0x3u) << PCR_PA_SHIFT)
+/* Cast to uint16_t is deliberate and provably lossless: the mask
+ * bounds n to 3, so the widest result is 3 << 10 = 0x0C00, well
+ * inside 16 bits. Stating it explicitly keeps -Wconversion usable as
+ * a standing check, so a genuine truncation bug would stand out. */
+#define PCR_AF(n)    ((uint16_t)(((uint32_t)(n) & 0x3u) << PCR_PA_SHIFT))
 
 /* Direct 16-bit register write - real, safe to use once the caller has
  * the correct PCR index (see the file-level warning above). */
